feat: concurrency policy triggered counter metric
Add a new metric called `cronworkflows_concurrencypolicy_triggered`
which counts the number of times a CronWorkflows `concurrencyPolicy`
has changed something.

The `concurency_policy` attribute can be:
* `Forbid`: Counts the number of times a workflow was not started
because one was already running
* `Replace`: Counts the number of times a workflow was terminated
because it was being replaced

This PR moves the call to increment `cronworkflows_triggered_total` so
that it doesn't count workflow runs suppressed by `Forbid`. This is
documented and the test updated to test it.

This PR also tidies some of the metrics documentation and language around
CronWorkflows.

Signed-off-by: Alan Clucas <alan@clucas.org>